auth-service 0.0.1

+++ auth-service/internal/auth/jwt.go
@@ -8,12 +8,12 @@ import(
 var jwtKey = []byte("super_secret_key")
 
 type Claims struct{
-	UserID int    `json:"user_id"`
+	UserID int64    `json:"user_id"`
     Email  string `json:"email"`
     jwt.RegisteredClaims
 }
 
-func GenerateToken(userID int, email string) (string, error){
+func GenerateToken(userID int64, email string) (string, error){
 	expirationTime := time.Now().Add(24 * time.Hour)
 
 	claims := &Claims{
